OneDrive/Desktop/imginvs/requirements_ui.txt
flask
pillow
numpy
tensorflow==2.20.0
